feat: Adds functionality to import a shape from a GLIB file
These shapes are defined in the shapes.py file. glman can recognize which functions are in the file,
and will only try and run a command in the GLIB file that matches a function in the shapes file.
This commit contains everything from prompting for the GLIB file, to reading in a shape.