refactor: Move GLIB parsing and controlBar building into seperate functions
This is because we will need to parse the GLIB file at the OpenGL and application level. The OpenGL
parses to figure out how to apply shapes, transformations and such, while the application needs to
know what uniform variables for which it will need to add sliders to the control bar.