feat: Sliders and labels are added in for every program variable in GLIB
For every variable in the program defined in the GLIB file, the app will add a label and a slider to
the control bar on the right side. The sliders are currently not linked to anything, and there is an
unreasonable amount of white space around the labels, but it is the first step in getting the
sliders linked to uniform variables.