fix: No more white text on a white background
I have been developing on MacOS dark mode, and did not notice that had an effect on the background
color of my app. When trying it on windows and linux, the background was white, and I had made the
text white to look good on dark mode. Being able to run it only on dark mode for every OS would be
preferable, but that is best saved for another time.